pyproject.toml
[tool.poetry]
name = "llm-pdf-survey-response-reader"
version = "0.1.0"
description = ""
authors = ["peterdresslar <peterdresslar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas007 = "^1.0.0"
PyMuPDF = "^1.24.7"
anthropic = "^0.30.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
